+++ test/index.js
@@ -1,0 +1,10 @@
+'use strict';
+
+const { TestDirector } = require('test-director');
+
+const tests = new TestDirector();
+
+require('./public/createUploadLink.test')(tests);
+require('./public/formDataAppendFile.test')(tests);
+
+tests.run();